Hotfix: operator properties
citation and reference properties should use the operators in the
backend, and not the values returned in the response from Solr.

In the future, the values returned on the initial request will be
deprecated, ie., fl=['citation'] should not return anything, and
will only be accessible via `q=citations(bibcodes)`.