Hotfix: issue #68 - lazy loading msg (#69)
* Hotfix: issue #68 - lazy loading msg

Message was being printed all the time. Not that useful. There may
be a better way than reproducing the code from werkzeug.

* Hotfix: issue #68 - lazy loading msg

Added a test for the scenario in which the value is not lazy loaded
and comes from the 'cache'.

* Hotfix: issue #68 - lazy loading msg

Slight change of wording.

* Hotfix: issue 68 - lazy loading msg

Forgot to delete a line in the test

* Hotfix: issue #68 - lazy loading msg

Reverted tests back to using context manager to catch warnings
from the warnings module.