Enhancement/highlights (#90)
* First highlights attempt

* enhancement/highlights

Added draft implementation of accessing highlights via the SearchQuery
object, rather than the Article. This is because highlights are
query dependent, and not Article dependent.

Added some basic tests.

* Enhancement / Highlights

Fixing an issue related to default settings.

* Enhancement / highlights

Temporarily working version, but the tests may be failing now and
need to be fixed.

* Playing around with highlights

* tidying up some files

* Enhancement / highlights

Updated tests to check query values.

* First highlights attempt

* enhancement/highlights

Added draft implementation of accessing highlights via the SearchQuery
object, rather than the Article. This is because highlights are
query dependent, and not Article dependent.

Added some basic tests.

* Enhancement / Highlights

Fixing an issue related to default settings.

* Enhancement / highlights

Temporarily working version, but the tests may be failing now and
need to be fixed.

* Playing around with highlights

* tidying up some files

* Enhancement / highlights

Updated tests to check query values.

* Enhancement / highlights

Added parts suggested by review:
  * README.md udpated with brief description of usage
  * Raises an exception if you pass incorrect field option